[Feat] QR 생성시 출석체크
