+++ Gongbu/View/DetailView/DetailQR/QRGenerator.swift
@@ -11,50 +11,20 @@ import Firebase
 
 struct QRGenerateView: View {
     
-    let roomID: String
+    let roomId: String
     @ObservedObject var studyModel = RankingModel()
     
     var body: some View {
         let userID:String = Auth.auth().currentUser!.uid
         VStack {
-            Image(uiImage: UIImage(data: returnData(str: self.roomID))!)
+            Image(uiImage: UIImage(data: returnData(str: self.roomId))!)
                 .resizable()
                 .frame(width: 150, height: 150)
             
         }
         .padding()
         .onAppear{
-            let study = studyModel.study
-            let now = Date()
-            // set db time
-            let setTime = study.time
-            let setDay = study.day
-            if setDay!.values.contains("\(now.toStringWeekday())요일") {
-                var db = Firestore.firestore()
-                let ref = db.collection("Study1").document(self.roomID)  // 룸 아이디 받으면 여기다가 넣어야됨ㅇㅅㅇ
-                let setDayKey:String = setDay!.someKey(forValue: "\(now.toStringWeekday())요일") ?? ""
-                let preparedTime = setTime![setDayKey]
-                let result = calculateTime(time: preparedTime!) - calculateTime(time: now.toStringTime())
-                print(result)
-                
-                //--------------
-                // 출석 완료
-                //--------------
-                if result >= 0 {
-                    ref.updateData([
-                        "currentAtt.\(userID)" : 5
-                    ])
-                }
-                
-                //--------------
-                // 지각
-                //--------------
-                else if result < 0 && result >= -10 {
-                    ref.updateData([
-                        "currentAtt.\(userID)" : 1
-                    ])
-                }
-            }
+            
         }
     }
